prolly wont do much here

+++ hw2/serial.cpp
@@ -10,12 +10,12 @@ std::vector<std::vector<std::vector<int>>> cells;
 
 // TODO this may change if our cells end up not being square after parallelism
 int get_cell_x(double size, double x) {
-    return ((num_cells - 1) * std::min((x / (size - cellSize)), 1.0));
+    return (int) ((num_cells - 1) * std::min((x / (size - cellSize)), 1.0));
 }
 
 // TODO this may change if our cells end up not being square after parallelism
 int get_cell_y(double size, double y) {
-    return ((num_cells-1) * std::min((y / (size - cellSize)), 1.0));
+    return (int) ((num_cells-1) * std::min((y / (size - cellSize)), 1.0));
 }
 
 // Apply the force from neighbor to particle
